refactor: extract the composer cluster into @endo/spaces-util; exclude it from the typecheck
Second slice of moving @endo/chat's view components into @endo/spaces-util.
Moves chat-bar's full import closure (23 files) via git mv into
packages/spaces-util/src/: chat-bar-component, send-form, command-selector,
inline-command-form, message-picker, blob-viewer, define-form, eval-form,
endow-modal, form-builder, help-modal, debugger-panel, command-executor,
command-registry, platform-keys, inline-define, inline-eval,
petname-path(s)-autocomplete, heat-bar, composite-heat-engine, heat-engine,
browser-tree. Confine helpers repointed off chat's setup-preact-container bundle
to @endo/preact-container/preact directly; chat.js + a few stay-in-chat files
repoint to @endo/spaces-util. chatBarComponent($parent, powers, options) entry +
the #chat-bar template are unchanged.

Type-check policy: @endo/spaces-util is now the shared chat-UI home — the
type-clean primitives plus the view components, which carry the same untyped-E()
code that was never type-checked in chat. Add packages/spaces-util to the root
tsconfig.json exclude (it was already in typedoc.json's), exactly as chat is
excluded, instead of casting ~26 untyped-E() errors across the moved UI (and
re-paying that tax on every future move). The shared primitives stay
type-checked transitively wherever a type-checked package (space-channel)
imports them; the view components are imported only by the excluded chat. Full
root typecheck is 0 errors.

vite build resolves all new imports; spaces-util + chat lint 0; 306
composer-cluster tests pass. No closure escape; browser-tree harden intact.
