+++ packages/chat/chat.js
@@ -12,6 +12,7 @@ import { channelComponent } from '@endo/space-channel/channel-component.js';
 import { forumComponent } from '@endo/space-channel/forum-component.js';
 import { microblogComponent } from '@endo/space-channel/microblog-component.js';
 import { valueComponent } from '@endo/spaces-util/value-component.js';
+import { chatBarComponent } from '@endo/spaces-util/chat-bar-component.js';
 import { createChannelHeader } from './channel-header.js';
 // Confined outliner host wrapper (the strangler-fig swap retired the imperative
 // `@endo/space-channel/outliner-component.js`; this is now the live outliner).
@@ -19,7 +20,6 @@ import { outlinerComponent } from './outliner-component.js';
 import { inboxComponent } from './inbox-component.js';
 import { inventoryComponent } from './inventory-component.js';
 import { channelListComponent } from './channel-list.js';
-import { chatBarComponent } from './chat-bar-component.js';
 import { createSpacesGutter } from './spaces-gutter.js';
 import { inventoryGraphComponent } from './inventory-graph-component.js';
 import { whylipComponent } from './whylip-component.js';
