+++ packages/chat/test/component/eval-form.test.js
@@ -46,7 +46,7 @@ register(
 // Dynamically imported AFTER the loader is registered so eval-form's
 // `import { createMonacoEditor } from '@endo/monaco-wrapper'` resolves to the
 // stub.
-const { createEvalForm } = await import('../../eval-form.js');
+const { createEvalForm } = await import('@endo/spaces-util/eval-form.js');
 
 const { document: testDocument } = createDOM();
