+++ packages/chat/test/component/petname-path-autocomplete-confined.test.js
@@ -5,9 +5,9 @@ import '@endo/init/debug.js';
 
 import test from 'ava';
 import { E } from '@endo/far';
+import { petNamePathAutocomplete } from '@endo/spaces-util/petname-path-autocomplete.js';
 import { createDOM, tick } from '../helpers/dom-setup.js';
 import { makeMockPowers } from '../helpers/mock-powers.js';
-import { petNamePathAutocomplete } from '../../petname-path-autocomplete.js';
 
 // Confined-conversion coverage for petname-path-autocomplete: the dropdown body
 // is now rendered through `renderConfined` into the host-provided `$menu`. These
@@ -31,6 +31,10 @@ const installRafShim = () => {
 /**
  * Poll until `predicate()` is true (or a timeout elapses). Copied from
  * inbox-shell.test.js: a fixed delay races on slower CI runners.
+ * @param predicate
+ * @param root0
+ * @param root0.timeout
+ * @param root0.step
  */
 const waitFor = async (predicate, { timeout = 3000, step = 20 } = {}) => {
   const start = Date.now();
