+++ packages/spaces-util/package.json
@@ -33,7 +33,30 @@
     "./chime.js": "./src/chime.js",
     "./locator.js": "./src/locator.js",
     "./package.json": "./package.json",
-    "./token-autocomplete.js": "./src/token-autocomplete.js"
+    "./token-autocomplete.js": "./src/token-autocomplete.js",
+    "./chat-bar-component.js": "./src/chat-bar-component.js",
+    "./send-form.js": "./src/send-form.js",
+    "./command-selector.js": "./src/command-selector.js",
+    "./inline-command-form.js": "./src/inline-command-form.js",
+    "./message-picker.js": "./src/message-picker.js",
+    "./blob-viewer.js": "./src/blob-viewer.js",
+    "./define-form.js": "./src/define-form.js",
+    "./eval-form.js": "./src/eval-form.js",
+    "./endow-modal.js": "./src/endow-modal.js",
+    "./form-builder.js": "./src/form-builder.js",
+    "./help-modal.js": "./src/help-modal.js",
+    "./debugger-panel.js": "./src/debugger-panel.js",
+    "./command-executor.js": "./src/command-executor.js",
+    "./command-registry.js": "./src/command-registry.js",
+    "./platform-keys.js": "./src/platform-keys.js",
+    "./inline-define.js": "./src/inline-define.js",
+    "./inline-eval.js": "./src/inline-eval.js",
+    "./petname-path-autocomplete.js": "./src/petname-path-autocomplete.js",
+    "./petname-paths-autocomplete.js": "./src/petname-paths-autocomplete.js",
+    "./heat-bar.js": "./src/heat-bar.js",
+    "./composite-heat-engine.js": "./src/composite-heat-engine.js",
+    "./heat-engine.js": "./src/heat-engine.js",
+    "./browser-tree.js": "./src/browser-tree.js"
   },
   "scripts": {
     "build": "exit 0",
@@ -43,11 +66,15 @@
   },
   "dependencies": {
     "@endo/daemon": "workspace:^",
+    "@endo/eventual-send": "workspace:^",
+    "@endo/exo": "workspace:^",
+    "@endo/exo-stream": "workspace:^",
     "@endo/far": "workspace:^",
     "@endo/harden": "workspace:^",
     "@endo/markmdown": "workspace:^",
     "@endo/monaco-wrapper": "workspace:^",
     "@endo/pass-style": "workspace:^",
+    "@endo/patterns": "workspace:^",
     "@endo/preact-container": "workspace:^",
     "marked": "^17.0.5",
     "preact": "^10.29.2"
