+++ packages/spaces-util/src/browser-tree.js
