+++ packages/spaces-util/src/chat-bar-component.js
@@ -7,14 +7,9 @@ import harden from '@endo/harden';
 import { E } from '@endo/far';
 import { iterateReader } from '@endo/exo-stream/iterate-reader.js';
 
-import {
-  Fragment,
-  h,
-  renderConfined,
-  unmount,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { Fragment, h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined, unmount } from '@endo/preact-container/renderer';
 
 import { sendFormComponent } from './send-form.js';
 import { commandSelectorComponent } from './command-selector.js';
