+++ packages/spaces-util/src/command-registry.js
