+++ packages/spaces-util/src/composite-heat-engine.js
