+++ packages/spaces-util/src/heat-engine.js
