+++ packages/spaces-util/src/inline-define.js
@@ -2,13 +2,9 @@
 
 import harden from '@endo/harden';
 
-import {
-  h,
-  renderConfined,
-  unmount,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined, unmount } from '@endo/preact-container/renderer';
 
 // Inline define input, migrated from imperative `innerHTML` DOM to a confined
 // Preact component rendered through a single `renderConfined`. The exported
