+++ packages/spaces-util/src/inline-eval.js
@@ -6,13 +6,9 @@
 
 import harden from '@endo/harden';
 
-import {
-  h,
-  renderConfined,
-  unmount,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined, unmount } from '@endo/preact-container/renderer';
 
 import { petNamePathAutocomplete } from './petname-path-autocomplete.js';
