+++ packages/spaces-util/src/message-picker.js
@@ -2,12 +2,9 @@
 
 import harden from '@endo/harden';
 
-import {
-  h,
-  renderConfined,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined } from '@endo/preact-container/renderer';
 
 // Message picker, migrated from imperative DOM to a confined Preact component
 // rendered through a single `renderConfined`.
