+++ packages/spaces-util/src/petname-path-autocomplete.js
@@ -6,14 +6,9 @@
 
 import harden from '@endo/harden';
 
-import {
-  Fragment,
-  h,
-  renderConfined,
-  unmount,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { Fragment, h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined, unmount } from '@endo/preact-container/renderer';
 
 // TODO: The autocomplete dropdown menu may not appear visually inside the eval
 // form modal despite the component working correctly (suggestions are fetched
