+++ packages/spaces-util/src/petname-paths-autocomplete.js
@@ -6,14 +6,9 @@
 
 import harden from '@endo/harden';
 
-import {
-  Fragment,
-  h,
-  renderConfined,
-  unmount,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { Fragment, h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined, unmount } from '@endo/preact-container/renderer';
 
 // Multi pet-name path autocomplete, migrated from imperative `innerHTML`/
 // `createElement` DOM to a confined Preact component rendered through a single
