+++ packages/spaces-util/src/send-form.js
@@ -6,14 +6,10 @@
 
 import harden from '@endo/harden';
 
-import { tokenAutocompleteComponent } from '@endo/spaces-util/token-autocomplete.js';
-import {
-  h,
-  renderConfined,
-  unmount,
-  useEffect,
-  useState,
-} from './setup-preact-container.js';
+import { h } from 'preact';
+import { useEffect, useState } from 'preact/hooks';
+import { renderConfined, unmount } from '@endo/preact-container/renderer';
+import { tokenAutocompleteComponent } from './token-autocomplete.js';
 
 import { makeLiveHeatEngine } from './heat-engine.js';
 import { makeCompositeHeatEngine } from './composite-heat-engine.js';
